usb-export: always unbind from current driver
Previously, usb-export ensured that the device was bound to usbip-host
by checking whether the device was already bound to it, and, if not,
unbind from the other driver and binding to usbip-host. usb-export then
checks whether the usbip_status for the device is 1 (available) and
aborts if otherwise.
This means that if the device is already bound to usbip-host, and
usbip_status is either 2 (used) or 3 (error), usb-export will not
succeed.

This change makes sure the current driver always gets unbound, even if
it is usbip-host, before (re)binding to usbip-host. It should hence
ensure that usb-export can succeed even if usbip_status was 2 or 3.

Previously, if the VM to which the USB device got attached managed to
cause an error in usbip-host, it was not possible to use the usual tools
to connect again, or to another VM: because usbip_status was not 2
(used) anymore, cleanup would be called so that the tools viewed the
device as not connected anymore, but attempting to attach the device to
anything again would fail due to the reason described above. The user
would then need to manually unbind the usbip-host driver from the device
prior to trying again. This change fixes this.