Clear cache when Qubes() object is getting destroyed
Commit a505ce7 "Send events even when device is attached/detached by
alternative method" introduced a local cache of usb devices. This cache
is valid only as long as Qubes() instance is. Normally it isn't an issue
(Qubes() object lifetime is bound to the qubesd process lifetime), but
during tests the object is created multiple times and cached objects
leaks between tests. Avoid this by clearing the cache whenever
Qubes().close() is called (and sends 'qubes-close' event).