Keep the process running to not close qrexec connection prematurely
Generally the actual USB packets are sent by the kernel into the socket,
so a process on either of the qrexec connection side isn't strictly
necessary. But keep both of them running, to not let the qrexec close
the connection early.

This is related to a qrexec code cleanup, which previously didn't closed
connection in some cases. Now it does it consistently, so usb-import
script can't rely on this bug anymore.