Add mechanism to reset device just before attaching
Some device require reset after interacting with previous driver. When
device was used with usbip-host before, it took care of that when
detaching from previous target (VM). But if it was attached to some
other driver in sys-usb before, this needs to be done manually.

To limit possible impact, make this opt-in. There are two methods to opt
into this feature:
1. Enable it for all devices using (adjust sys-usb name if applicable):
   qvm-service --enable sys-usb usb-reset-on-attach
2. Enable for specific device by setting QUBES_USB_RESET udev property.

QubesOS/qubes-issues#8953